feat(#131): Move PositionPublisher to perception
