+++ code/acting/launch/acting.launch
@@ -19,17 +19,12 @@
     </node>
 
     <node pkg="acting" type="DummyTrajectoryPublisher.py" name="DummyTrajectoryPublisher" output="screen">
-          <param name="control_loop_rate" value="1" />
-          <param name="role_name" value="$(arg role_name)" />
+        <param name="control_loop_rate" value="1" />
+        <param name="role_name" value="$(arg role_name)" />
     </node>
 
-    <node pkg="acting" type="DummyTrajectorySub.py" name="DummyTrajectorySub" output="screen">
-          <param name="control_loop_rate" value="1" />
-          <param name="role_name" value="$(arg role_name)" />
-    </node>
-
-        <node pkg="acting" type="PositionPublisher.py" name="PositionPublisher" output="screen">
-          <param name="control_loop_rate" value="1" />
-          <param name="role_name" value="$(arg role_name)" />
-    </node>
+    <!--node pkg="acting" type="DummyTrajectorySub.py" name="DummyTrajectorySub" output="screen">
+        <param name="control_loop_rate" value="1" />
+        <param name="role_name" value="$(arg role_name)" />
+    </node-->
 </launch>
